server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.5.0
python-dotenv==1.0.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.13
email-validator==2.2.0